requirements.txt
numpy==1.17.0
pillow==5.4.1
absl-py